LayoutTests/media/video-controls.html

<video controls></video>
<p>Test 'controls' attribute<p>
<script src=../http/tests/media/media-file.js></script>
<script src=../http/tests/media/video-test.js></script>
<script>
    testExpected("video.controls", null, '!=');
    waitForEvent('canplaythrough', function () {
        testExpected("video.controls", null, '!=');
        run("video.removeAttribute('controls')");
        testExpected("video.controls", false, '==');
        run("video.controls = true");
        testExpected("video.getAttribute('controls')", "");
        run("video.controls = false");
        testExpected("video.controls", false, '==');
        run("video.setAttribute('controls','controls')");
        testExpected("video.controls", null, '!=');
        endTest();
    } );
    video.src = findMediaFile("video", "content/test");
</script>
